+++ src/app/_components/local-status-box/book-mark-box.tsx
@@ -1,40 +1,40 @@
 "use client";
 
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { useState } from "react";
 import { Separator } from "@/components/ui/separator";
 import Link from "next/link";
-import { removeSearchHistory } from "@/lib/searchHistoryFunc";
+import { ScrollArea } from "@/components/ui/scroll-area";
+import { removeBookmarkId } from "@/lib/book-mark-func";
+import { useState } from "react";
 
-function getSearchHistories(): string[] {
-  const searchHistoryString = localStorage.getItem("searchHistory");
-  return searchHistoryString ? JSON.parse(searchHistoryString) : [];
+function getBookmarks(): string[] {
+  const bookmarkString = localStorage.getItem("bookmark");
+  return bookmarkString ? JSON.parse(bookmarkString) : [];
 }
 
-export default function SearchHistoryBox() {
-  const [searchHistoryArr, setSearchHistoryArr] =
-    useState(getSearchHistories());
+export default function BookMarkBox() {
+  const [bookmarkArr, setBookmarkArr] = useState(getBookmarks());
+
   return (
-    <ScrollArea className="h-29 w-full rounded-md border">
+    <ScrollArea className="h-22 w-full rounded-md border">
       <div className="p-4">
         <h4 className="mb-4 flex justify-center font-medium leading-none">
-          검색기록
+          즐겨찾기
         </h4>
-        {searchHistoryArr.map((searchHistoryId, idx) => {
-          const [id, key] = searchHistoryId.split("#");
+        {bookmarkArr.map((bookmark, idx) => {
+          const [id, key] = bookmark.split("#");
           return (
             <div key={idx}>
               <Separator className="my-1" />
               <div className="flex justify-between items-center w-full">
                 <Link href={`/summoner-page?id=${id}&tag=${key}`}>
                   <div className=" text-xs w-full h-10 flex justify-center items-center">
-                    {searchHistoryId}
+                    {bookmark}
                   </div>
                 </Link>
                 <button
                   onClick={() => {
-                    removeSearchHistory(searchHistoryId);
-                    setSearchHistoryArr(getSearchHistories());
+                    removeBookmarkId(bookmark);
+                    setBookmarkArr(getBookmarks());
                   }}
                   className="text-xs"
                 >
