+++ src/app/_components/search-bar/index.tsx
@@ -2,12 +2,12 @@
 
 import { useRouter } from "next/navigation";
 import React, { useEffect, useRef, useState, useTransition } from "react";
-import createSubmitHandler from "@/app/_components/createSubmitHandler";
+import createSubmitHandler from "@/app/_components/create-submit-handler";
 import axios from "axios";
 import Index from "@/app/_components/local-status-box";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import ReadingGlassSvgIcon from "@/components/ui/readingGlassSvgIcon";
+import ReadingGlassSvgIcon from "@/components/ui/reading-glass-svg-icon";
 import RecommendedNicknameList from "@/app/_components/search-bar/recommended-nickname-list";
 
 export default function SearchBar() {
