+++ src/app/summoner-page/_components/error-page.tsx
