+++ src/app/summoner-page/_components/user-info-box/book-mark-button.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import BookMarkStar from "@/components/ui/bookMarkStar";
+import BookMarkStar from "@/components/ui/book-mark-star";
 import { useEffect, useState } from "react";
 import {
   addBookmarkId,
   isLocalStorageBookmarked,
   removeBookmarkId,
-} from "@/lib/bookMarkFunc";
+} from "@/lib/book-mark-func";
 
 interface BookMarkStarButtonProps {
   id: string;
