+++ src/components/ui/reading-glass-svg-icon.tsx
