+++ src/lib/search-history-func.ts
